Deleted unneeded repository interfaces and files.  Moved still needed but misplaced files to their appropriate packages to prep for conversion to session bean files.
