+++ src/main/java/com/portfolio/onlineStore/service/ReportsService.java
@@ -1,5 +1,5 @@
 package com.portfolio.onlineStore.service;
 
-public class ReportsServiceImpl {
+public class ReportsService {
 
 }
